pyproject.toml
[tool.poetry]
name = "wimshow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
websockets = "^9"
tqdm = "^4"
pypng = "^0.0.20"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
opencv-python = "^4.5"
simplejpeg = "^1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
